Only call hide() method in the popup menu when it's visible.
When the menu is opened and an item is selected, the menu will be hidden.  When the menu button is clicked again, handleBlur() event handler will be called to hide the menu again (at this time, focus will be moved from the menu to the button), right before the showWithPosition() method is called.  The extra call to hide() method will record the lastHide_ timestamp, which is pretty close to the time showWithPosition() method is called.  In this method, wasRecentlyHide() will almost always return false (threshold is 150ms at this time), and prevent the menu from being displayed.

R=pupius
DELTA=4  (4 added, 0 deleted, 0 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=31