Automated g4 rollback.
*** Reason for rollback ***

This is breaking all urls with & in them by first html
escaping the & to & and then escaping the semicolon giving
&%3B.  The result is that the browser tries to navigate to
foo&;bar instead of foo&bar.


*** Original change description ***

Deal with URLs that contain unencoded semicolons by enclosing the
url of a refresh meta tag in additional single quotes. On IE, the
semicolon was treated as a separator in the content attribute,
causing the URL to be cut off.


Tested this with Chrome 8, IE8, FF3.6, FF3, FF2.

R=pupius
DELTA=13  (0 added, 12 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=287