Prevent dragStart from being missed when dragging close to the element border.
The current code fires a dragStart event after the distance from the initial click exceeds a certain threshold. However the distance is no tracked after the mouse leaves the clicked element so the event may never be fired if the element is dragged close to its border.

If that situation occurs the listeners are not unregistered so even if the mouse is release, the next time it hovers the initial drag element it will finally trigger the dragStart event.

This bug seems to be affecting most closure apps that use drag and drop, including the closure demos, or the Gmail inbox.

This CL implements a simple fix by firing the dragStart event when the mouse leaves the dragged element, regardless of the traveled distance.

R=pallosp,eae,nicksantos


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2215