Firefox has a slightly different way of handling the requestAnimationFrame method (even though it was the original implementation), which is not always compatible with the W3C draft specs used by Webkit. Also there is a possible bug caused by different threading handling.
Im not sure why this is so, but I think it's a result of agressive
optimalization that may separate the data objects if the post-compiler don't
see the need for running the same objects / threads as base for the callbacks.
But when using the event-handler, it will definetly break that optimalization
(e.g. the call-graph is too big).

R=pupius
DELTA=72  (63 added, 2 deleted, 7 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3498